main: pull apart fsqueue and fsmem
