+++ src/main/fsmem.h
@@ -1,5 +1,5 @@
-#ifndef MAIN_4_H
-#define MAIN_4_H
+#ifndef _FSMEM_H
+#define _FSMEM_H
 
 #include "common.h"
