+++ src/main/fsqueue_2.c
@@ -1,0 +1,3 @@
+#include "fsqueue.h"
+
+INCLUDE_ASM("asm/main/nonmatchings/fsqueue_2", fsQueueUpdateRead);
