--- src/main/main_3.c
@@ -1,3 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("asm/main/nonmatchings/main_3", fsQueueUpdateRead);
